requirements.txt
openai-whisper
yt_dlp
h5py
pandas
python-dotenv
tables
langchain
langchain-openai
click
questionary
rich
